requirements.txt
streamlit
google-generativeai
load_dotenv